:bug: bug fixes in if condition of consecutiveRuns
The behaviour of && and & is different in if condition: & and &&
indicate logical AND and | and || indicate logical OR. The shorter
form performs elementwise comparisons in much the same way as
arithmetic operators. The longer form evaluates left to right
examining only the first element of each vector.
Evaluation proceeds only until the result is determined.
The longer form is appropriate for programming control-flow
and typically preferred in if clauses.


Former-commit-id: 7b7efdaf5a303c43f73f499ce131c615121c747a